Change pymultio versioning
Originally setuptools_scm was used for versioning the pymultio wheel.
I'm switching that to using the VERSION file -- to have it aligned
with multio itself as well the multiolib python wheel.

This is needed in particular for our custom `.devYMD` releases.

The `_version.py` is, for simplicity, changed to dynamically read
via `importlib.metadata`, instead of static hardcoded version at
build time.